+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - environment variables are displayed on their own tab
 - infrastructure list is not longer full width
 - users can see infrastructure variables as well as environment variables for a given infrastructure
+- users can create and update infrastructure variables
 
 [unreleased]: TODO
 [keep a changelog 1.0.0]: https://keepachangelog.com/en/1.0.0/